Refactor PRHandler and PRDashboard into view_pr.py
